upload stem function

+++ uploadFile.php
@@ -13,6 +13,9 @@ function uploadFile($filetempname,$leadExcel)
     $uploadfile=$filePath.date("y-m-d h-m-s").".xlsx";//upload file name    
   
     $result=move_uploaded_file($filetempname,$uploadfile);//move to upload folder  
+    $msg = "fail";
+    $count = 0;
+    $active = "N";
 
     if($result) //if upload successed, then read and insert data  
     {   
@@ -39,8 +42,10 @@ function uploadFile($filetempname,$leadExcel)
     	    $DepartmentID = $objPHPExcel->getActiveSheet()->getCell("H".$j)->getValue();
     	    $sql = "INSERT INTO People VALUES('".$UID."','".$DepartmentID."','".$LastName."','".$FirstName."','".$FullName."','".$Email."','".$Phone."','".$Title."')";
     	    //echo $sql;
-        if(mysql_query($sql,$conn))
-           $msg = "success";      	}
+        // if(mysql_query($sql,$conn))
+        //    $msg = "success";    
+        if(mysql_query($sql,$conn)) $count++;   	
+       }
     }
     if($leadExcel == 'school'){
       for($j=2;$j<=$highestRow;$j++)
@@ -50,8 +55,10 @@ function uploadFile($filetempname,$leadExcel)
           $OrgID = $objPHPExcel->getActiveSheet()->getCell("C".$j)->getValue();
           $sql = "INSERT INTO School_Division VALUES('".$SchoolDivisionID."','".$SchoolDivisionName."','".$OrgID."')";
           //echo $sql;
-        if(mysql_query($sql,$conn))
-           $msg = "success";        }
+        // if(mysql_query($sql,$conn))
+        //    $msg = "success";        
+        if(mysql_query($sql,$conn)) $count++; 
+       }
     }
     if($leadExcel == 'department'){
       for($j=2;$j<=$highestRow;$j++)
@@ -61,8 +68,9 @@ function uploadFile($filetempname,$leadExcel)
           $SchoolDivisionID = $objPHPExcel->getActiveSheet()->getCell("C".$j)->getValue();
           $sql = "INSERT INTO Department_Center VALUES('".$DepartmentID."','".$DepartmentName."','".$SchoolDivisionID."')";
           //echo $sql;
-        if(mysql_query($sql,$conn))
-           $msg = "success";  
+        // if(mysql_query($sql,$conn))
+        //    $msg = "success"; 
+        if(mysql_query($sql,$conn)) $count++;  
       }
     }
     if($leadExcel == 'fundingsource'){
@@ -73,10 +81,45 @@ function uploadFile($filetempname,$leadExcel)
           $ShortName = $objPHPExcel->getActiveSheet()->getCell("C".$j)->getValue();
           $sql = "INSERT INTO FundingSource VALUES('".$FundID."','".$FundName."','".$ShortName."')";
           //echo $sql;
-        if(mysql_query($sql,$conn))
-           $msg = "success";  
+        // if(mysql_query($sql,$conn))
+        //    $msg = "success"; 
+        if(mysql_query($sql,$conn)) $count++;  
       }
     }
+    if($leadExcel == 'stem'){
+      for($j=2;$j<=$highestRow;$j++)
+      {
+          $STEMID = $objPHPExcel->getActiveSheet()->getCell("A".$j)->getValue();
+          $UID = $objPHPExcel->getActiveSheet()->getCell("B".$j)->getValue();
+          if(empty($UID))
+            $UID='0';
+          $FundID = $objPHPExcel->getActiveSheet()->getCell("C".$j)->getValue();
+          if(empty($FundID))
+            $FundID='0';            
+          $GrantID = $objPHPExcel->getActiveSheet()->getCell("D".$j)->getValue();
+           if(empty($GrantID))
+            $GrantID='0';
+          $InitiativeName = $objPHPExcel->getActiveSheet()->getCell("E".$j)->getValue();
+          $Relationship = $objPHPExcel->getActiveSheet()->getCell("F".$j)->getValue();
+          $FundAmount = $objPHPExcel->getActiveSheet()->getCell("G".$j)->getValue();
+          $StartDate = PHPExcel_Style_NumberFormat::toFormattedString($objPHPExcel->getActiveSheet()->getCell("H".$j)->getValue(), 'YYYY-MM-DD');
+          $EndDate = PHPExcel_Style_NumberFormat::toFormattedString($objPHPExcel->getActiveSheet()->getCell("I".$j)->getValue(), 'YYYY-MM-DD');
+          $Evaluator = $objPHPExcel->getActiveSheet()->getCell("J".$j)->getValue();
+          $ProjectURL = $objPHPExcel->getActiveSheet()->getCell("K".$j)->getValue();
+
+          $sql = "INSERT INTO STEM VALUES('".$STEMID."','".$UID."','".$FundID."','".$GrantID."','".$InitiativeName."',
+                                          '".$Relationship."','".$FundAmount."','".$StartDate."','".$EndDate."',
+                                          '".$Evaluator."','".$ProjectURL."','".$active."')";
+          //echo $sql;
+        if(mysql_query($sql,$conn)) $count++; 
+      }
+    }
+
+    if($count == $highestRow)
+      $msg = "success"; 
+    else
+      $msg = "incomplete";
+
        unlink ($uploadfile); //delete uploaded file 
        mysql_close($conn); 
     }else{
